src/requirements.txt
cloudshell-automation-api>=8.2,<8.3
cloudshell-core>=2.2,<2.3
cloudshell-shell-core>=3.1,<3.2
cloudshell-traffic-virtual>=1.0,<1.1
cloudshell-networking-devices>=2.5,<2.6
requests==2.18.4
